# src/utils/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "version": "1.0.0",
      "title": "API Restful",
      "description": "Routes disponibles du projet"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "components": {
        "securitySchemes": {
          "jwt": {
            "type": "http",
            "scheme": "bearer",
            "in": "header",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [{
        "jwt": []
      }],
    "paths": {
...
        "/api/user/{id}": {
            "delete": {
                "tags": ["Users"],
                "description": "Supprimer un utilisateur",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'utilisateur à supprimer",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            }
        },
...
    },
...
}